Fixes non public instances Getter and Setter, class refactoring
